moved #confirmation_required? from LicensesHandlers to LicensesFetchers
It's basically looking for a special file in the license archive. Thus it's
more similar to the other fetcher methods and can profit from the cached
license archive in LicensesFetchers.
